Changing the unit test to mapdl_test
